fix: updated inputcomponents to match new validation flow
