+++ packages/core/src/components/question/input-types/text/TextInput.tsx
@@ -1,63 +1,10 @@
 "use client";
-import React, { CSSProperties, InputHTMLAttributes } from "react";
-import { ChangeEvent, useState } from "react";
-import classNames from "classnames";
-import styles from "./TextInput.module.css";
-import { useQuickForm } from "../../../../state/QuickFormContext";
+import React from "react";
 import { TextProperties } from "../../../../model/InputType";
 import { InputComponentType, registerInputComponent } from "../../../../services/defaults/DefaultInputTypeResolver";
-import { useFocusableQuestion } from "../../../../hooks/useFocusableQuestion";
-import { QuestionModel } from "../../../../model";
 import { textInputSchema } from "./TextInputSchema";
-import { useFocusOutHandler } from "../../../../hooks/useFocusOutHandler";
+import { BaseInputComponent } from "../baseinput/BaseInputComponent";
 
-export const BaseInputComponent = ({ questionModel, className, style, type }: { type: InputHTMLAttributes<HTMLInputElement>["type"], questionModel: QuestionModel, className?: string, style?: CSSProperties }) => {
-
-    const [text, setText] = useState<string>(questionModel!.output);
-    const ref = useFocusableQuestion<HTMLInputElement>(questionModel.logicalName);
-    const { answerQuestion, validateAndAnswerQuestion } = useQuickForm();
-
-    const resize = () => {
-        const input = ref.current;
-        if (!input)
-            return;
-
-        const oldvalue = input.value;
-
-        if (!oldvalue || oldvalue === '')
-            input.value = input.placeholder;
-
-        const isOverflowed = input.scrollWidth > input.clientWidth;
-        input.value = oldvalue;
-        if (isOverflowed) {
-            var style = window.getComputedStyle(input, null).getPropertyValue('font-size');
-            input.style.fontSize = (parseFloat(style) - 1) + "px";
-            resize();
-        }
-    }
-
-    /* @pks - I added a quick hook to attempt adressing iOS focus issues - please test if this works as expected */
-    useFocusOutHandler(ref, async (value) => await validateAndAnswerQuestion(questionModel.logicalName, value));
-
-    const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
-        if (event.target.value === "")
-            questionModel.errorMsg = "";
-        setText(() => event.target.value);
-        answerQuestion(questionModel.logicalName, event.target.value);
-        resize();
-    }
-
-    return (
-        <input style={style}
-            ref={ref}
-            type={type}
-            className={classNames(styles.input__text, className)}
-            placeholder={questionModel.placeholder}
-            value={text}
-            onChange={handleChange}
-        />
-    );
-}
 export const TextInput: InputComponentType<TextProperties> = (props) => {
     return <BaseInputComponent type="text" {...props} />
 }
